AIM-144 refactor: change enum object to string
